+++ integrations/facebook-pixel/lib/index.js
@@ -24,8 +24,10 @@ var FacebookPixel = module.exports = integration('Facebook Pixel')
   .option('valueIdentifier', 'value')
   .option('initWithExistingTraits', false)
   .option('traverse', false)
+  .option('automaticConfiguration', true)
   .mapping('standardEvents')
   .mapping('legacyEvents')
+  .mapping('contentTypes')
   .tag('<script src="//connect.facebook.net/en_US/fbevents.js">');
 
 /**
@@ -52,6 +54,9 @@ FacebookPixel.prototype.initialize = function() {
   window.fbq.version = '2.0';
   window.fbq.queue = [];
   this.load(this.ready);
+  if (!this.options.automaticConfiguration) {
+    window.fbq('set', 'autoConfig', false, this.options.pixelId);
+  }
   if (this.options.initWithExistingTraits) {
     var traits = formatTraits(this.analytics);
     window.fbq('init', this.options.pixelId, traits);
@@ -183,15 +188,15 @@ FacebookPixel.prototype.productListViewed = function(track) {
   // If no products have been defined, fallback on legacy behavior.
   // Facebook documents the content_type parameter decision here: https://developers.facebook.com/docs/facebook-pixel/api-reference
   if (contentIds.length) {
-    contentType = 'product';
+    contentType = ['product'];
   } else {
     contentIds.push(track.category() || '');
-    contentType = 'product_group';
+    contentType = ['product_group'];
   }
 
   window.fbq('track', 'ViewContent', {
     content_ids: contentIds,
-    content_type: contentType
+    content_type: this.mappedContentTypesOrDefault(track.category(), contentType), 
   });
 
   // fall through for mapped legacy conversions
@@ -213,7 +218,7 @@ FacebookPixel.prototype.productListViewed = function(track) {
 FacebookPixel.prototype.productViewed = function(track) {
   window.fbq('track', 'ViewContent', {
     content_ids: [track.productId() || track.id() || track.sku() || ''],
-    content_type: 'product',
+    content_type: this.mappedContentTypesOrDefault(track.category(), ['product']),
     content_name: track.name() || '',
     content_category: track.category() || '',
     currency: track.currency(),
@@ -239,7 +244,7 @@ FacebookPixel.prototype.productViewed = function(track) {
 FacebookPixel.prototype.productAdded = function(track) {
   window.fbq('track', 'AddToCart', {
     content_ids: [track.productId() || track.id() || track.sku() || ''],
-    content_type: 'product',
+    content_type: this.mappedContentTypesOrDefault(track.category(), ['product']),
     content_name: track.name() || '',
     content_category: track.category() || '',
     currency: track.currency(),
@@ -263,18 +268,27 @@ FacebookPixel.prototype.productAdded = function(track) {
  */
 
 FacebookPixel.prototype.orderCompleted = function(track) {
+  var products = track.products() || [];
+
   var content_ids = foldl(function(acc, product) {
     var item = new Track({ properties: product });
     var key = item.productId() || item.id() || item.sku();
     if (key) acc.push(key);
     return acc;
-  }, [], track.products() || []);
+  }, [], products);
 
   var revenue = formatRevenue(track.revenue());
 
+  // Order completed doesn't have a top-level category spec'd.
+  // Let's default to the category of the first product. - @gabriel
+  var contentType = ['product_group'];
+  if (products.length) {
+    contentType = this.mappedContentTypesOrDefault(products[0].category, contentType);
+  }
+
   window.fbq('track', 'Purchase', {
     content_ids: content_ids,
-    content_type: 'product',
+    content_type: contentType,
     currency: track.currency(),
     value: revenue
   });
@@ -288,6 +302,22 @@ FacebookPixel.prototype.orderCompleted = function(track) {
   }, this.legacyEvents(track.event()));
 };
 
+/**
+ * mappedContentTypesOrDefault returns an array of mapped content types for
+ * the category - or returns the defaul value.
+ * @param {Facade.Track} track 
+ * @param {Array} def 
+ */
+FacebookPixel.prototype.mappedContentTypesOrDefault = function(category, def) {
+  if (!category) return def
+
+  var mapped = this.contentTypes(category);
+  if (mapped.length) {
+    return mapped
+  } 
+  
+  return def
+}
 
 /**
  * Get Revenue Formatted Correctly for FB.