+++ integrations/facebook-pixel/HISTORY.md
@@ -1,3 +1,10 @@
+2.5.1 / 2018-09-10
+==================
+
+  * Makes content type customizable for Product List Viewed, Product Viewed, Product Added, and
+  Order Completed events.
+  * Adds a setting to disable Automatic Configuration.
+
 2.5.0 / 2018-06-22
 ==================